Fixed sensors not updating
Fixed that the sensors wont update.

In the code i'm checking if the latest update from the sensor is older then 10 mintus. This is to stop updating the sensor even in the night when no new information is on the portal. But the first time the sensor are loaded it always needs to gather the information of course ;-)